# Map.Api/wwwroot/swagger/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Map.API",
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/User/{userId}/email": {
      "patch": {
        "tags": [
          "User"
        ],
        "summary": "Change user email",
        "operationId": "UpdateUserMailAsyncPATCH",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "Id of user",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "description": "UpdateUserMailDto",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserMailDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserMailDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserMailDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TripDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Server Error",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/User/Contact": {
      "patch": {
        "tags": [
          "User"
        ],
        "operationId": "ContactMailAsyncPATCH",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MailDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/MailDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/MailDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Server Error",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "nullable": true
          },
          "message": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "MailDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "nullable": true
          },
          "subject": {
            "type": "string",
            "nullable": true
          },
          "body": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "TripDto": {
        "type": "object",
        "properties": {
          "tripId": {
            "type": "string",
            "description": "Id of the trip.",
            "format": "uuid"
          },
          "userId": {
            "type": "string",
            "description": "id of the user who created the trip.",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "description": "Name of the trip.",
            "nullable": true
          },
          "description": {
            "type": "string",
            "description": "Description of the trip.",
            "nullable": true
          },
          "startDate": {
            "type": "string",
            "description": "Start date of the trip.",
            "format": "date"
          },
          "endDate": {
            "type": "string",
            "description": "End date of the trip.",
            "format": "date"
          },
          "backgroundPicturePath": {
            "type": "string",
            "description": "Path of the background picture.",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "UpdateUserMailDto": {
        "type": "object",
        "properties": {
          "mail": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "JWT authorization header using the bearer scheme. <br />\r\nEnter your token in the text input below. <br />\r\nExample: '12345abcdef' <br />\r\nNB: A cookie will be added in order to remember the authentication during the session,\r\nif you try to logout using swagger logout feature this might not work.",
        "scheme": "Bearer"
      }
    }
  }
}