# Map.Domain/Models/Email/MailDto.cs
namespace Map.Domain.Models.EmailDto;
public class MailDto
{
    public string Email { get; set; }
    public string Subject { get; set; }
    public string Body { get; set; }
    public string Name { get; set; }
}